requirements.txt
streamlit
requests
pandas
libsql-client
python-dotenv